convert input to int32:

+++ multipath.py
@@ -10,7 +10,7 @@
 import sys
 from scipy.io.wavfile import read as readwav
 from scipy.io.wavfile import write as writewav
-from numpy import power, int16
+from numpy import power, int16, int32
 from os import mkdir
 import string
 
@@ -39,6 +39,9 @@ def main():
 	print(f'opened {sys.argv[1]}, sample rate {input_sample_rate}')
 
 	sample_count = len(input_audio)
+
+	# convert input audio to int32
+	input_audio = input_audio.astype(int32)
 
 	# calculate multipath sample delay
 	start_milliseconds = float(sys.argv[2]) * input_sample_rate / 1000
